- changed DPI value to float

+++ trunk/android/android/src/NativeInterface.java
@@ -114,16 +114,16 @@ public class NativeInterface
 		return result;
 	}
 	
-	public static int getDisplayDpi()
+	public static float getDisplayDpi()
 	{
 		// hardcoded exceptions for known devices that return wrong DPI
 		if (Build.MANUFACTURER.equals("HTC") && Build.MODEL.equals("HTC One X"))
 		{
-			return 312;
+			return 312.0;
 		}
 		DisplayMetrics metrics = new DisplayMetrics();
 		NativeInterface.Activity.getWindowManager().getDefaultDisplay().getMetrics(metrics);
-		return (int)Math.sqrt((metrics.xdpi * metrics.xdpi + metrics.ydpi * metrics.ydpi) / 2.0);
+		return Math.sqrt((metrics.xdpi * metrics.xdpi + metrics.ydpi * metrics.ydpi) / 2.0);
 	}
 	
 	public static String getOsVersion()
